pyproject.toml
[tool.poetry]
name = "hikeguard-bot"
version = "0.1.0"
description = "Telegram anti-spam bot for Ukrainian hiking community"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
aiogram = ">=3.25.0"
aiosqlite = ">=0.22.0"
google-genai = ">=1.65.0"
pydantic-settings = ">=2.13.0"
python-dotenv = ">=1.0.0"
structlog = ">=25.5.0"
aiohttp = ">=3.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "DTZ", "SIM"]

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)"
]
